requirements.txt
fastapi
uvicorn
kafka-python
apify-client
python-dotenv
boto3
pydantic
aiokafka
